main.rs: Use byte slices instead of char slices
`char` is a special type to represent a "unicode scalar value" which is
not accurate for most of the TA.  Simply using bytes everywhere and
parsing UTF-8 strings where necessary makes the whole data slice easier
to deal with.

At the same time, borrow rules help save a whole lot of iterations and
copies.